add fused_multi_transformer to incubate
